Working priorities + crosshair in front
git-svn-id: https://legesmotus.svn.sourceforge.net/svnroot/legesmotus/legesmotus/trunk@69 a5a5d5d7-22d1-44d1-b37a-67d3d99fcfd7

+++ client/Sprite.hpp
@@ -52,7 +52,7 @@ public:
 
 	double	get_x() const;
 	double	get_y() const;
-	double	get_priority() const;
+	int	get_priority() const;
 
 	double	get_scale_x() const;
 	double	get_scale_y() const;
